remove old design files and mvc controllers
