+++ src/main/java/com/project/mizawra/dao/JournalRepository.java
@@ -6,7 +6,6 @@ import java.util.List;
 import java.util.UUID;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 
 public interface JournalRepository extends JpaRepository<Journal, UUID> {
     List<Journal> findAllByOwner(Client owner, Pageable pageable);
