+++ src/main/java/com/project/mizawra/service/JournalService.java
@@ -1,6 +1,5 @@
 package com.project.mizawra.service;
 
-import com.project.mizawra.models.Client;
 import com.project.mizawra.models.Journal;
 import com.project.mizawra.models.dto.JournalDto;
 import java.util.List;
@@ -12,5 +11,5 @@ public interface JournalService {
     List<Journal> getJournals(int page);
     Long getPageCount();
     Journal save(JournalDto journalDto) throws Exception;
-    long countByOwner(Client owner);
+    Long countJournalsForAuthenticatedUser();
 }
