--- src/main/resources/fragment/navbar.html
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html lang="en"
-      xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta charset="UTF-8">
-    <link href="/css/global.css" rel="stylesheet"/>
-</head>
-<body>
-<nav th:fragment="navbar" id="navbar" class="navbar">
-    <div class="logo">MIZAWRA</div>
-    <div class="menu">
-        <a sec:authorize="!isAuthenticated()" href="/login">login</a>
-        <a sec:authorize="!isAuthenticated()" href="/home">about</a>
-
-        <a sec:authorize="isAuthenticated()" href="/home">home</a>
-        <a sec:authorize="isAuthenticated()" href="/settings">settings</a>
-        <a sec:authorize="isAuthenticated()" href="/login">logout</a>
-    </div>
-</nav>
-</body>
-</html>
