--- src/main/resources/static/css/almostDone.css
@@ -1,11 +1,0 @@
-h1 {
-  margin: 15px 0;
-}
-
-.container {
-  padding: 42px 168px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  text-align: center;
-}
