--- src/main/resources/static/css/global.css
@@ -1,87 +1,0 @@
-@import url("https://fonts.googleapis.com/css?family=Open+Sans:400,600,300");
-@import url("https://fonts.googleapis.com/css?family=Chivo:900");
-
-body {
-    background-color: #0F1010;
-    color: white;
-    margin: 0;
-    font-family: 'Open Sans';
-}
-
-body a {
-    text-decoration: none;
-    color: #3A6BA4;
-}
-
-.navbar {
-    display: flex;
-    justify-content: space-between;
-    padding: 42px 84px;
-}
-
-.navbar a {
-    color: white;
-    font-weight: 800;
-    font-size: 24px;
-    line-height: 33px;
-}
-
-.navbar .menu {
-    display: flex;
-    gap: 60px;
-}
-
-.logo {
-    font-family: 'Chivo';
-    font-size: 28px;
-}
-
-.button {
-    width: 344px;
-    height: 56px;
-    color: white;
-    background-color: #232223;
-    box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.4);
-    border: none;
-    border-radius: 10px;
-}
-
-.color-grey {
-    color: #EBECF1;
-    opacity: 0.4;
-}
-
-.d-none {
-    display: none !important;
-}
-
-input {
-    width: -webkit-fill-available;
-    padding: 15px;
-    margin: 15px 100px;
-    background: #232223;
-    color: white;
-    border: none;
-    border-radius: 10px;
-}
-
-h1 {
-    font-size: 36px;
-    margin: 0;
-}
-
-h2 {
-    font-size: 24px;
-    margin: 0;
-}
-
-h3 {
-   font-weight: 700;
-   font-size: 20px;
-   line-height: 30px;
-   margin: 0;
-}
-
-p {
-    margin: 0;
-}
