--- src/main/resources/static/css/home.css
@@ -1,58 +1,0 @@
-.home {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 30px;
-}
-
-.home h1 {
-    margin: 0;
-}
-
-.user-bar {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    gap: 30px;
-}
-
-.user-info {
-    display: grid;
-    gap: 5px;
-    width: 550px;
-    align-self: baseline;
-    margin: 20px 0;
-}
-
-.link {
-    font-weight: 500;
-    font-size: 20px;
-    line-height: 30px;
-
-    width: min-content;
-    text-align: center;
-    color: white;
-}
-
-.link .mizawra {
-    font-family: Chivo;
-}
-
-.journal-container {
-    display: grid;
-    grid-template-columns: 1fr 1fr 1fr;
-    gap: 60px;
-    text-align-last: center;
-    margin-bottom: 40px;
-}
-
-.journal-container a {
-    font-weight: 800;
-    font-size: 24px;
-    line-height: 33px;
-    color: white;
-
-    padding: 30px;
-    box-shadow: 2px 4px 8px rgba(0, 0, 0, 0.4);
-    border-radius: 10px;
-}
