--- src/main/resources/static/css/login.css
@@ -1,36 +1,0 @@
-.form {
-    display: grid;
-    justify-items: center;
-    gap: 20px;
-
-    height: fit-content;
-
-    position: absolute;
-    top: 20%;
-    left: 0;
-    right: 0;
-    bottom: 0;
-}
-
-.form.register {
-    top: 10%;
-}
-
-.form .footer {
-    display: flex;
-    gap: 5px;
-}
-
-.input-form {
-    width: 560px;
-    padding: 36px 0;
-    text-align: center;
-
-    box-sizing: border-box;
-    border: 3px solid #252626;
-    border-radius: 10px;
-}
-
-h1 {
-    margin: 15px 0;
-}
