--- src/main/resources/static/css/writing.css
@@ -1,91 +1,0 @@
-body {
-    height: 70vh;
-}
-
-.container {
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-    position: relative;
-    text-align: center;
-
-    margin: 0 84px;
-    padding: 35px 150px;
-
-    background: #1A1C1C;
-    border: 2px solid #484E4E;
-    border-radius: 10px;
-}
-
-.time-container {
-    position: absolute;
-    top: 40px;
-    right: 40px;
-
-    cursor: pointer;
-}
-
-.prompt-container {
-    padding: 20px 0;
-}
-
-.textarea-container {
-    flex: 1;
-    display: grid;
-    grid-template-columns: 1fr;
-    grid-template-rows: 70px 1fr;
-}
-
-.textarea-container .title {
-    font-weight: 800;
-    font-size: 36px;
-    line-height: 49px;
-
-    color: #EBECF1;
-    width: 100%;
-    margin: 0;
-    padding: 0;
-
-    background: none;
-    border: none;
-    overflow: auto;
-    outline: none;
-
-    -webkit-box-shadow: none;
-    -moz-box-shadow: none;
-    box-shadow: none;
-}
-
-textarea {
-    font-weight: 400;
-    font-size: 20px;
-    line-height: 30px;
-
-    color: #EBECF1;
-
-    width: 100%;
-
-    background: none;
-    border: none;
-    overflow: auto;
-    outline: none;
-
-    -webkit-box-shadow: none;
-    -moz-box-shadow: none;
-    box-shadow: none;
-
-    resize: none;
-}
-
-button {
-    font-weight: 800;
-    font-size: 24px;
-    line-height: 29px;
-
-    color: #FFFFFF;
-    padding: 35px 65px;
-    background: #3AA451;
-    border-radius: 10px;
-
-    cursor: pointer;
-}
