--- src/main/resources/static/js/forgetPassword.js
@@ -1,10 +1,0 @@
-document.addEventListener("DOMContentLoaded", init);
-
-function init() {
-    document.getElementById('confirm').addEventListener('click', ()=>{
-        const email = document.getElementById('email').value;
-        $.post("/forgetPassword", {email: email}, function(data) {
-            window.location.href = "/login?message=" + data.message;
-        });
-    });
-}
