--- src/main/resources/templates/home.html
@@ -1,35 +1,0 @@
-<!DOCTYPE html>
-<html lang="en"
-      xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:th="http://www.thymeleaf.org">
-<head>
-    <title>MiZaWra</title>
-    <meta charset="UTF-8">
-
-    <link href="/css/global.css" rel="stylesheet"/>
-    <link href="/css/home.css" rel="stylesheet">
-
-    <script src="/js/home.js" type="module"></script>
-    <script src="https://code.jquery.com/jquery-3.5.1.min.js" integrity="sha256-9/aliU8dGd2tb6OSsuzixeV4y/faTqgFtohetphbbj0=" crossorigin="anonymous"></script>
-</head>
-<body>
-    <nav th:replace="~{../fragment/navbar :: navbar}"></nav>
-    <div class="home">
-        <div class="user-bar">
-            <img height="128px" width="128px" src="/images/profile.jpg">
-            <div class="user-info">
-                <h2 th:text="${client.getFirstName() + ' ' + client.getLastName()}"></h2>
-                <p th:text="#{home.defaultDescription}" class="color-grey">We plan to make you have the ability to share your texts with other writers. Therefore, in the future this text will be your bio and you can pick whatever picture you want. For now: Tell yourself why you use MiZaWra (in less than 300 characters).</p>
-            </div>
-            <span class="link" th:text="${countJournals} + ' posts'">33 posts</span>
-            <a class="link" href="/about">about <div class="mizawra">MIZAWRA</div></a>
-        </div>
-        <h1 th:text="#{home.h1}">create a new narrative</h1>
-        <div class="journal-container">
-            <a href="/promptedJournal" style="background: #3A6BA4;" th:text="#{home.promptedJournal}">prompted journal</a>
-            <a href="/write" style="background: #3AA451;" th:text="#{home.freeJournal}">free journal</a>
-            <a href="/story" style="background: #8062FF;" th:text="#{home.story}">story</a>
-        </div>
-    </div>
-</body>
-</html>
