--- src/main/resources/templates/login/almostDone.html
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html lang="en"
-      xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:th="http://www.thymeleaf.org">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>MiZaWra</title>
-
-  <link href="/css/global.css" rel="stylesheet"/>
-  <link href="/css/almostDone.css" rel="stylesheet"/>
-
-</head>
-<body>
-<nav th:replace="~{../fragment/navbar :: navbar}"></nav>
-  <div class="container">
-    <h1 th:text="#{almostDone.h1}"></h1>
-    <p th:text="#{almostDone.p}"></p>
-    <p><br><a href="mailto:mizawrawriting@gmail.com">mizawrawriting@gmail.com</a></p>
-  </div>
-</form>
-</body>
-</html>
