--- src/main/resources/templates/login/changePassword.html
@@ -1,23 +1,0 @@
-<html
-        xmlns="http://www.w3.org/1999/xhtml"
-        xmlns:th="http://www.thymeleaf.org">
-<head>
-    <link href="/css/global.css" rel="stylesheet"/>
-    <link href="/css/login.css" rel="stylesheet"/>
-
-    <script type="module" src="/js/changePassword.js"></script>
-    <script src="https://code.jquery.com/jquery-3.5.1.min.js" integrity="sha256-9/aliU8dGd2tb6OSsuzixeV4y/faTqgFtohetphbbj0=" crossorigin="anonymous"></script>
-</head>
-<body>
-    <nav th:replace="~{../fragment/navbar :: navbar}"></nav>
-    <form class="form">
-        <div class="input-form">
-            <h1 th:text="#{changePassword.h1}">Choose your new password</h1>
-            <input id="password" name="password" type="password" th:placeholder="#{generic.inputs.password}">
-            <input id="matchPassword" name="matchPassword" type="password" th:placeholder="#{changePassword.matching}">
-        </div>
-        <button th:text="#{generic.inputs.confirm}" class="button" id="confirm" type="submit"></button>
-        <a th:href="@{/login}">Login</a>
-    </form>
-</body>
-</html>
