--- src/main/resources/templates/login/forgetPassword.html
@@ -1,21 +1,0 @@
-<html
-        xmlns="http://www.w3.org/1999/xhtml"
-        xmlns:th="http://www.thymeleaf.org">
-<head>
-    <link href="/css/global.css" rel="stylesheet"/>
-    <link href="/css/login.css" rel="stylesheet"/>
-    <script type="module" src="/js/forgetPassword.js"></script>
-    <script src="https://code.jquery.com/jquery-3.5.1.min.js" integrity="sha256-9/aliU8dGd2tb6OSsuzixeV4y/faTqgFtohetphbbj0=" crossorigin="anonymous"></script>
-</head>
-<body>
-    <nav th:replace="~{../fragment/navbar :: navbar}"></nav>
-    <div class="form">
-        <div class="input-form">
-            <h1 th:text="#{forgetPassword.h1}">Reset your password</h1>
-            <input id="email" name="email" type="email" th:placeholder="#{generic.inputs.email}">
-        </div>
-        <button th:text="#{generic.inputs.confirm}" class="button" id="confirm"></button>
-        <p th:text="#{forgetPassword.remember}"><a href="/login">Login</a></p>
-    </div>
-</body>
-</html>
