--- src/main/resources/templates/login/login.html
@@ -1,29 +1,0 @@
-<!DOCTYPE html>
-<html lang="en"
-      xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:th="http://www.thymeleaf.org">
-  <head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>MiZaWra</title>
-
-    <link href="/css/global.css" rel="stylesheet"/>
-    <link href="/css/login.css" rel="stylesheet"/>
-  </head>
-  <body>
-    <nav th:replace="~{../fragment/navbar :: navbar}"></nav>
-    <form class="form" action="login" method="post">
-      <div class="input-form">
-        <h1 th:text="#{login.h1}"></h1>
-        <input type="email" id="username" name="username" th:placeholder="#{generic.inputs.email}"><br>
-        <input type="password" id="password" name="password" th:placeholder="#{generic.inputs.password}"><br><br>
-        <a href="/register/forgetPassword" th:text="#{login.forgotPW}"></a>
-      </div>
-      <input class="button" name="submit" type="submit" value="Login" />
-      <div class="footer">
-        <p th:text="#{login.noAccount}"></p>
-        <a href="/register" th:text="#{login.signUp}"></a>
-      </div>
-    </form>
-  </body>
-</html>
