--- src/main/resources/templates/login/register.html
@@ -1,31 +1,0 @@
-<!DOCTYPE html>
-<html xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:th="http://www.thymeleaf.org"
-      lang="en">
-    <head>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <title>MiZaWra</title>
-
-        <link href="/css/global.css" rel="stylesheet"/>
-        <link href="/css/login.css" rel="stylesheet"/>
-    </head>
-    <body>
-        <nav th:replace="~{../fragment/navbar :: navbar}"></nav>
-        <form class="form register" action="register" th:object="${client}" method="post">
-            <div class="input-form">
-                <h1 th:text="#{register.h1}">Registration</h1>
-                <input type="text" id="firstName" name="firstName" th:placeholder="#{generic.inputs.firstName}"><br>
-                <input type="text" id="lastName" name="lastName" th:placeholder="#{generic.inputs.lastName}"><br>
-                <input type="email" id="email" name="email" th:placeholder="#{generic.inputs.email}"><br>
-                <input type="password" id="password" name="password" th:placeholder="#{generic.inputs.password}"><br><br>
-            </div>
-
-            <input class="button" name="submit" type="submit" th:value="#{register.createAccountButton}" />
-            <div class="footer">
-                <p th:text="#{register.alreadyAccount}"></p>
-                <a href="/login">Login</a>
-            </div>
-        </form>
-    </body>
-</html>
