--- src/main/resources/templates/message.html
@@ -1,8 +1,0 @@
-<html
-        xmlns="http://www.w3.org/1999/xhtml"
-        xmlns:th="http://www.thymeleaf.org">
-<body>
-    <h1 th:text="${message}">Error Message</h1>
-    <a th:href="@{/login}">Login</a>
-</body>
-</html>
