--- src/main/resources/templates/promptedJournal.html
@@ -1,41 +1,0 @@
-<!DOCTYPE html>
-<html lang="en"
-      xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:th="http://www.thymeleaf.org">
-<head>
-    <title>MiZaWra</title>
-    <meta charset="UTF-8">
-
-    <link href="/css/global.css" rel="stylesheet"/>
-    <link href="/css/prompted.css" rel="stylesheet">
-
-    <script src="/js/prompt.js" type="module"></script>
-</head>
-<body>
-    <nav th:replace="~{../fragment/navbar :: navbar}"></nav>
-    <div class="container">
-        <div class="prompt-header">
-            <h1 th:text="#{prompted.h1}">what do you want to write about?</h1>
-            <p th:text="#{promptedJournal.h1Description}">Each sub-mode will give you a question to answer in your journal entry. </p>
-        </div>
-        <div class="prompt-container">
-            <div id="philosophical" class="prompt">
-                <h2 th:text="#{promptedJournal.subPhilosophical}">philosophical prompt</h2>
-                <p>Wanna try to answer the big questions today? Go for this sub-mode!</p>
-            </div>
-            <div id="positive" class="prompt">
-                <h2 th:text="#{promptedJournal.subPositive}">positive prompt</h2>
-                <p>Feel like you need something uplifting today? This sub-mode will try to make you think about something positive. </p>
-            </div>
-            <div id="neutral" class="prompt">
-                <h2 th:text="#{promptedJournal.subNeutral}">neutral prompt</h2>
-                <p>Just in the mood for some reflection on who you are and what you believe? Try this sub-mode!</p>
-            </div>
-            <div id="negative" class="prompt">
-                <h2 th:text="#{promptedJournal.subNegative}">negative prompt</h2>
-                <p>Feel like tackling a  difficult personal moment today?</p>
-            </div>
-        </div>
-    </div>
-</body>
-</html>
