--- src/main/resources/templates/writing.html
@@ -1,35 +1,0 @@
-<!DOCTYPE html>
-<html lang="en"
-      xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:th="http://www.thymeleaf.org">
-<head>
-    <title>MiZaWra</title>
-    <meta charset="UTF-8">
-
-    <link href="/css/global.css" rel="stylesheet"/>
-    <link href="/css/writing.css" rel="stylesheet">
-
-    <script src="/js/writing.js" type="module"></script>
-    <script src="https://code.jquery.com/jquery-3.5.1.min.js" integrity="sha256-9/aliU8dGd2tb6OSsuzixeV4y/faTqgFtohetphbbj0=" crossorigin="anonymous"></script>
-</head>
-<body>
-    <nav th:replace="~{../fragment/navbar :: navbar}"></nav>
-    <div th:id="${journalId}" class="container">
-        <div th:if="${mode == null && title== null && body == null}" class="time-container">
-            <h1>15:00</h1>
-            <p class="color-grey" th:text="#{writing.timerSubtext}">click to pause/start</p>
-        </div>
-        <div th:if="${prompt != null}" class="prompt-container">
-            <p class="color-grey">your prompt</p>
-            <h3 id="prompt" th:attr="prompt-id=${prompt.getId()}" th:text="${prompt.getPrompt()}">What can you hope for?</h3>
-        </div>
-        <div class="textarea-container">
-            <input id="title" class="title" name="title" type="text" th:value="${title}" th:placeholder="#{writing.placeholderTitle}">
-            <textarea id="journal" name="journal" th:text="${body}" th:placeholder="#{writing.placeholderWritingArea}" autofocus></textarea>
-        </div>
-        <div>
-            <button id="save" th:text="#{writing.exitButton}">save & exit</button>
-        </div>
-    </div>
-</body>
-</html>
